fix: assign single-tagset windows to their matching output and cache render config lookups

+++ src/kwm/layout/grid.zig
@@ -33,7 +33,7 @@ pub fn arrange(self: *const Self, output: *Output) !void {
 
     if (windows.items.len == 0) return;
 
-    const col_num: i32 = @intFromFloat(@ceil(@sqrt(@as(f64, @floatFromInt(windows.items.len)))));
+    const col_num: i32 = @intFromFloat(@ceil(@sqrt(@as(f32, @floatFromInt(windows.items.len)))));
     const row_num: i32 = @intFromFloat(@ceil(@as(f32, @floatFromInt(windows.items.len)) / @as(f32, @floatFromInt(col_num))));
     const available_width, const available_height = blk: {
         const width = @max(0, output.exclusive_width() - 2 * self.outer_gap);
